backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
anthropic==0.39.0
openai==1.58.1
google-generativeai==0.8.3
requests==2.31.0
PyPDF2==3.0.1
pillow==10.1.0
python-multipart==0.0.6